+++ addon/mixins/limited-controller.js
@@ -1,24 +1,9 @@
 import Ember from 'ember';
 
 export default Ember.Mixin.create({
-  queryParams: ['lf', 'filter'],
-
-  lf: null,
+  queryParams: ['filter'],
 
   filter: null,
-
-  /**
-   * Update current limit function.
-   *
-   * @method updateLimitFunction
-   * @param {String} limitFunction New limit function.
-   */
-  updateLimitFunction: function(limitFunction) {
-    if (this.get('lf') !== limitFunction) {
-      // Changing lf value reloads route automatically.
-      this.set('lf', limitFunction);
-    }
-  },
 
   actions: {
     /**
@@ -29,7 +14,10 @@ export default Ember.Mixin.create({
      */
     filterByAnyMatch: function(pattern) {
       if (this.get('filter') !== pattern) {
-        this.set('filter', pattern);
+        this.setProperties({
+          filter: pattern,
+          page: 1
+        });
       }
     }
   }