+++ tests/dummy/app/models/ember-flexberry-dummy-comment.js
@@ -26,7 +26,7 @@ var Model = BaseModel.extend({
 
   // Model validation rules.
   validations: {
- }
+  }
 });
 
 // Edit form projection.