+++ tests/dummy/app/models/ember-flexberry-dummy-suggestion-type.js
@@ -63,4 +63,22 @@ Model.defineProjection('SuggestionTypeL', 'ember-flexberry-dummy-suggestion-type
   })
 });
 
+// Projection for lookup example on window customization.
+Model.defineProjection('CustomizeLookupWindowExampleView', 'ember-flexberry-dummy-suggestion-type', {
+  name: Proj.attr('Name'),
+  moderated: Proj.attr('Moderated')
+});
+
+// Projection for lookup with limit function example.
+Model.defineProjection('LookupWithLimitFunctionExampleView', 'ember-flexberry-dummy-suggestion-type', {
+  name: Proj.attr('Name'),
+  moderated: Proj.attr('Moderated')
+});
+
+// Projection for lookup in dropdown mode example.
+Model.defineProjection('DropDownLookupExampleView', 'ember-flexberry-dummy-suggestion-type', {
+  name: Proj.attr('Name'),
+  moderated: Proj.attr('Moderated')
+});
+
 export default Model;