+++ tests/dummy/app/serializers/ember-flexberry-dummy-suggestion.js
@@ -3,15 +3,16 @@ import ApplicationSerializer from './application';
 // TODO: ODataSerializer.extend
 export default ApplicationSerializer.extend({
   attrs: {
-  editor1: { serialize: 'odata-id', deserialize: 'records' },
-  type: { serialize: 'odata-id', deserialize: 'records' },
-  author: { serialize: 'odata-id', deserialize: 'records' },
-  files: { serialize: false, deserialize: 'records' },
-  comments: { serialize: false, deserialize: 'records' },
-  userVotes: { serialize: false, deserialize: 'records' }
-},
+    editor1: { serialize: 'odata-id', deserialize: 'records' },
+    type: { serialize: 'odata-id', deserialize: 'records' },
+    author: { serialize: 'odata-id', deserialize: 'records' },
+    files: { serialize: false, deserialize: 'records' },
+    comments: { serialize: false, deserialize: 'records' },
+    userVotes: { serialize: false, deserialize: 'records' }
+  },
+
   /**
-  * Field name where object identifier is kept.
-  */
+    Property name in which object identifier is kept.
+   */
   primaryKey: '__PrimaryKey'
 });